+++ Examples/Swift/GDPRExample/GDPRExample/ViewController.swift
@@ -10,7 +10,7 @@ import YandexMobileAds
 
 class ViewController: UIViewController {
     var adLoader: YMANativeAdLoader!
-    var lastBanner: YMANativeBannerView?
+    var bannerView: YMANativeBannerView?
     var gdprManager: GDPRUserConsentManager!
 
     override func viewDidLoad(){
@@ -34,7 +34,6 @@ class ViewController: UIViewController {
         navigationController?.pushViewController(settings, animated: true)
     }
 
-    
     private func showGDPRDialog() {
         // This is sample GDPR dialog which is used to demonstrate the GDPR user consent retrieval logic.
         // Please, do not use this dialog in production app.
@@ -44,54 +43,26 @@ class ViewController: UIViewController {
         self.present(dialog, animated: true, completion: nil)
     }
 
-    private func didLoadAd(ad : YMANativeGenericAd) {
-        let bannerView = YMANativeBannerView()
-        bannerView.ad = ad
-
-        lastBanner?.removeFromSuperview()
-        lastBanner = bannerView
-        self.view.addSubview(bannerView)
-
-        bannerView.translatesAutoresizingMaskIntoConstraints = false
-
-        if #available(iOS 11.0, *) {
-            configureLayoutAtBottomOfSafeArea(for: bannerView)
-        } else {
-            configureLayoutAtBottom(for: bannerView)
-        }
+    private func loadAd() {
+        adLoader.loadAd(with: nil)
     }
 
     private func configureLayoutAtBottom(for bannerView: UIView) {
-        let views = ["bannerView" : bannerView]
-        let horizontal = NSLayoutConstraint.constraints(withVisualFormat: "H:|-[bannerView]-|",
-                                                        options: [],
-                                                        metrics: nil,
-                                                        views: views)
-        let vertical = NSLayoutConstraint.constraints(withVisualFormat: "V:[bannerView]-|",
-                                                      options: [],
-                                                      metrics: nil,
-                                                      views: views)
-        self.view.addConstraints(horizontal)
-        self.view.addConstraints(vertical)
-    }
-
-    @available(iOS 11.0, *)
-    private func configureLayoutAtBottomOfSafeArea(for bannerView: UIView) {
-        let guide = self.view.safeAreaLayoutGuide
+        var guide = self.view.layoutMarginsGuide
+        if #available(iOS 11.0, *) {
+            guide = self.view.safeAreaLayoutGuide
+        }
+        bannerView.translatesAutoresizingMaskIntoConstraints = false
         let constraints = [bannerView.leadingAnchor.constraint(equalTo: guide.leadingAnchor),
                            bannerView.trailingAnchor.constraint(equalTo: guide.trailingAnchor),
                            bannerView.bottomAnchor.constraint(equalTo: guide.bottomAnchor)]
         NSLayoutConstraint.activate(constraints)
     }
-
-    private func loadAd() {
-        adLoader.loadAd(with: nil)
-    }
 }
 
 // MARK: - GDPRDialogDelegate
 
-extension ViewController: YMANativeAdLoaderDelegate {
+extension ViewController: GDPRDialogDelegate {
     func dialogDidDismiss(_ dialog : GDPRDialogViewController) {
         loadAd()
     }
@@ -99,20 +70,18 @@ extension ViewController: YMANativeAdLoaderDelegate {
 
 // MARK: - YMANativeAdLoaderDelegate
 
-extension ViewController: GDPRDialogDelegate {
-    func nativeAdLoader(_ loader: YMANativeAdLoader!, didLoad ad: YMANativeAppInstallAd) {
-        didLoadAd(ad: ad)
+extension ViewController: YMANativeAdLoaderDelegate {
+    func nativeAdLoader(_ loader: YMANativeAdLoader, didLoad ad: YMANativeAd) {
+        let bannerView = YMANativeBannerView()
+        bannerView.ad = ad
+
+        self.bannerView?.removeFromSuperview()
+        self.bannerView = bannerView
+        view.addSubview(bannerView)
+        configureLayoutAtBottom(for: bannerView)
     }
 
-    func nativeAdLoader(_ loader: YMANativeAdLoader!, didLoad ad: YMANativeContentAd) {
-        didLoadAd(ad: ad)
-    }
-
-    func nativeAdLoader(_ loader: YMANativeAdLoader!, didLoad ad: YMANativeImageAd) {
-        didLoadAd(ad: ad)
-    }
-
-    func nativeAdLoader(_ loader: YMANativeAdLoader!, didFailLoadingWithError error: Error) {
+    func nativeAdLoader(_ loader: YMANativeAdLoader, didFailLoadingWithError error: Error) {
         print("Native ad loading error: ", error)
     }
 }