+++ changelog/adapter/admob-mobileads/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 0.11.0
+
+#### Updated
+* Updated minimum supported Yandex Mobile Ads SDK version to 3.0.0
+
 ## Version 0.10.0
 
 #### Updated